refactor: Move RegisterViewModel to correct package
