+++ app/src/main/java/com/paris_2/san3a/di/uiModule.kt
@@ -1,7 +1,7 @@
 package com.paris_2.san3a.di
 
 import com.paris_2.san3a.presentation.screen.register.OTPRegisterViewModel
-import com.paris_2.san3a.presentation.screen.register.registerScreen.components.RegisterViewModel
+import com.paris_2.san3a.presentation.screen.register.registerScreen.RegisterViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
